add-service-create

+++ src/main/java/com/rental_api/ServiceBooking/Dto/Request/ServiceRequest.java
@@ -1,12 +1,11 @@
 package com.rental_api.ServiceBooking.Dto.Request;
 
 import lombok.Data;
-import java.math.BigDecimal;
 
 @Data
 public class ServiceRequest {
     private String name;
     private String description;
-    private BigDecimal price;
-    private Integer duration; // in minutes
+    private Double price;
+    private Integer duration; // duration in minutes
 }
